moved convert into epd lib
